chore(shared): modularize replacement methods
